# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "description": "API for PostAway Application",
        "title": "PostAway API"
    },
    "servers": [
        {
            "url": "http://localhost:3200"
        }
    ],
    "components": {
        "securitySchemes": {
            "JWT": {
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": ["Users"],
                "summary": "Get Users",
                "description": "User will get all the users",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": ["Users"],
                "summary": "Sign Up",
                "description": "User Signup to create an account",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": ["name", "email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/users/signin": {
            "post": {
                "tags": ["Users"],
                "summary": "Login",
                "description": "User Login to get the token",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": ["email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Incorrect Credentials"
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "tags": ["Posts"],
                "summary": "Get Posts",
                "description": "User will get all the posts",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": ["Posts"],
                "summary": "Add Post",
                "description": "User can add a new post",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "caption": {
                                        "type": "string"
                                    },
                                    "imageUrl": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/posts/caption": {
            "get": {
                "tags": ["Posts"],
                "summary": "Filter Posts by Caption",
                "description": "User will filter posts by caption",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "parameters": [
                    {
                        "name": "caption",
                        "in": "query",
                        "description": "Caption to filter posts",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/posts/userposts": {
            "get": {
                "tags": ["Posts"],
                "summary": "Get Posts by User",
                "description": "User will get all posts by a specific user",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/comments": {
            "get": {
                "tags": ["Comments"],
                "summary": "Get Comments",
                "description": "User will get all the comments",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": ["Comments"],
                "summary": "Add Comment",
                "description": "User can add a new comment",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "postId": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "required": ["postId", "content"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/comments/update": {
            "put": {
                "tags": ["Comments"],
                "summary": "Update Comment",
                "description": "User can update an existing comment",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "commentId": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "required": ["commentId", "content"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/comments/delete": {
            "delete": {
                "tags": ["Comments"],
                "summary": "Delete Comment",
                "description": "User can delete an existing comment",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "commentId": {
                                        "type": "string"
                                    }
                                },
                                "required": ["commentId"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/likes": {
            "get": {
                "tags": ["Likes"],
                "summary": "Get Likes",
                "description": "User will get all the likes",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/likes/toggle/": {
            "post": {
                "tags": ["Likes"],
                "summary": "Toggle Like",
                "description": "User can toggle like on a post",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "postId": {
                                        "type": "string"
                                    }
                                },
                                "required": ["postId"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/saves": {
            "get": {
                "tags": ["Saves"],
                "summary": "Get Saved Posts",
                "description": "User will get all saved posts",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/saves/toggle/": {
            "post": {
                "tags": ["Saves"],
                "summary": "Toggle Save",
                "description": "User can toggle save on a post",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "postId": {
                                        "type": "string"
                                    }
                                },
                                "required": ["postId"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    }
}